Volume of interest limited Gaussian 3d smoothing
Volume of interest limited Gaussian 3d smoothing for volumetric imaging. The intensity values outside the VOI are considered unknown and are not used in the filtering process.  These are also placed in my old website, which is still up, but I do not have write access to that site anymore.